adding batch of changes 5

+++ testing/cuda/gather.cu
@@ -24,7 +24,7 @@ void TestGatherDevice(ExecutionPolicy exec, const size_t n)
   thrust::host_vector<unsigned int> h_map = unittest::random_integers<unsigned int>(n);
   
   for(size_t i = 0; i < n; i++)
-    h_map[i] =  h_map[i] % source_size;
+    h_map[i] = h_map[i] % static_cast<unsigned int>(source_size);
   
   thrust::device_vector<unsigned int> d_map = h_map;
   
@@ -117,7 +117,7 @@ void TestGatherIfDevice(ExecutionPolicy exec, const size_t n)
   thrust::host_vector<unsigned int> h_map = unittest::random_integers<unsigned int>(n);
   
   for(size_t i = 0; i < n; i++)
-      h_map[i] = h_map[i] % source_size;
+      h_map[i] = h_map[i] % static_cast<unsigned int>(source_size);
   
   thrust::device_vector<unsigned int> d_map = h_map;
